backend/requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt